// samples/react-sp-tinymce/src/webparts/tinymceEditor/model/FieldType.ts
export enum FieldType {
  Text = 'Text',
  Note = 'Note',
  Lookup = 'Lookup',
  LookupMulti = 'LookupMulti',
  Choice = 'Choice',
  MultiChoice = 'MultiChoice',
  Number = 'Number',
  Currency = 'Currency',
  DateTime = 'DateTime',
  Boolean = 'Boolean',
  User = 'User',
  UserMulti = 'UserMulti',
  URL = 'URL',
  File = 'File',
  Computed = 'Computed',
  TaxonomyFieldType = 'TaxonomyFieldType',
  TaxonomyFieldTypeMulti = 'TaxonomyFieldTypeMulti',
  Location = 'Location',
  Thumbnail = 'Thumbnail',
  Attachments = 'Attachments',
  AllDayEvent = 'AllDayEvent'
}